requirements.txt
django
djangorestframework
psycopg2
django-cors-headers
django-secure
django-sslserver